refactor: 70% update for mvc and prep
