+++ src/front/pages/game.html
@@ -7,22 +7,5 @@
     <link rel="stylesheet" href="../style.css" />
     <title>Game Time</title>
   </head>
-  <body class="game-page">
-    <header class="header">
-      <div class="game-options">
-        <div class="game-src">
-          <button class="game-src_btn" data-src="server">Server</button>
-          <button class="game-src_btn" data-src="storage">LocalStorage</button>
-        </div>
-      </div>
-    </header>
-    <h1 class="header">It's just a game, boy!</h1>
-    <a href="/">Welcome back</a>
-    <div class="game">
-      <button class="game-start-btn btn">Start Game</button>
-    </div>
-    <div class="audio">
-      <audio src="/audio" controls>Music</audio>
-    </div>
-  </body>
+  <body class="game-page"></body>
 </html>
